Chapter 5
1. Clean up code
2. Create folder structure for app
3. Make different files for each component.
4. Create a config file
5. Use all types of import and export
6. Create a Search Box in your App
7. Use useState to create a variable and bind it to the input box
8. Try to make your search bar work
