+++ index.css
@@ -33,3 +33,7 @@
     display: flex;
     flex-wrap: wrap;
 }
+
+.search-input{
+    margin: 10px;
+}
