+++ index.html
@@ -12,9 +12,6 @@
 <body>
     <!-- <h1>hello</h1> -->
     <div id="root">Not Rendered</div>
-
-    <!-- <script crossorigin src="https://unpkg.com/react@18/umd/react.development.js"></script> -->
-    <!-- <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.development.js"></script> -->
     <script>
         // const heading = document.createElement('h1');
         // heading.innerHTML = 'Hello';
@@ -23,7 +20,7 @@
         // root.appendChild(heading);
     </script>
     <!-- <script type="module" src="app.js"></script> -->
-    <script type="module" src="app.js"></script>
+    <script type="module" src="src/app.js"></script>
 </body>
 
 </html>
