+++ src/app.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import ReactDOM from "react-dom/client";
+import Header from './components/Header';
+import Body from './components/Body';
+import Footer from './components/Footer';
+
+const AppLayout = () => {
+    return (
+        <>
+            <Header />
+            <Body />
+            <Footer />
+        </>
+    );
+}
+
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(<AppLayout />)
+
+/** 
+   Header
+    - Logo(Title)
+    - Nav Items
+    - Cart
+   Body
+    - Search bar
+    - RestaurantList
+        - RestaurantCard (Multiple)
+            - Image
+            - Name
+            - Rating 
+            - Cusines
+    Footer
+        - Links
+        - Copyright
+*/
