+++ src/basic2.js
