+++ src/components/Footer.js
@@ -1,0 +1,5 @@
+const Footer = () => {
+    return <h4>Footer</h4>
+}
+
+export default Footer;
