+++ src/constant.js
@@ -1,53 +1,6 @@
-import React from 'react';
-import ReactDOM from "react-dom/client";
+export const IMG_CDN_URL = 'https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/';
 
-/** 
-   Header
-    - Logo(Title)
-    - Nav Items
-    - Cart
-   Body
-    - Search bar
-    - RestaurantList
-        - RestaurantCard (Multiple)
-            - Image
-            - Name
-            - Rating 
-            - Cusines
-    Footer
-        - Links
-        - Copyright
-*/
-
-const Title = () => {
-    return (
-        <a href='/'>
-            <img
-                className='logo'
-                alt='logo'
-                src='https://yt3.ggpht.com/ytc/AMLnZu_EC-ECXAxRAixWGEfMsE1rdSoetBHyxmLNdtCB=s900-c-k-c0x00ffffff-no-rj'
-            />
-        </a>
-    )
-}
-
-const Header = () => {
-    return (
-        <div className='header'>
-            <Title />
-            <div className='nav-items'>
-                <ul>
-                    <li>Home</li>
-                    <li>About</li>
-                    <li>Contact</li>
-                    <li>Cart</li>
-                </ul>
-            </div>
-        </div>
-    )
-}
-
-const restaurantList = [
+export const restaurantList = [
     {
         "type": "restaurant",
         "data": {
@@ -1420,47 +1373,3 @@ const restaurantList = [
         "subtype": "basic"
     },
 ]
-
-const RestaurantCard = ({ name, cuisines, cloudinaryImageId, lastMileTravelString }) => {
-    return (
-        <div className='card'>
-            <img
-                src={
-                    'https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/' +
-                    cloudinaryImageId
-                }
-            />
-            <h2>{name}</h2>
-            <h3>{cuisines.join(", ")}</h3>
-            <h4>{lastMileTravelString} minutes</h4>
-        </div>
-    )
-}
-
-const Body = () => {
-    return (
-        <div className='restaurant-list'>
-            {restaurantList.map((restaurant) => {
-                return <RestaurantCard {...restaurant.data} key={restaurant.data.id} />
-            })}
-        </div>
-    )
-}
-
-const Footer = () => {
-    return <h4>Footer</h4>
-}
-
-const AppLayout = () => {
-    return (
-        <>
-            <Header />
-            <Body />
-            <Footer />
-        </>
-    )
-}
-
-
-const root = ReactDOM.createRoot(document.getElementById('root'))
-root.render(<AppLayout />)
